chore: remove and ignore yarn.lock
It looks like this file was generated at the beginning of the project,
but a fresh clone of the repo and running `yarn install` resulted in a
large update to the yarn.lock file. It seems like the package-lock.json
file is kept up-to-date, and the yarn.lock file can be ignored from the
repo.